openssl vendored

+++ src/cli/actions/mod.rs
@@ -2,5 +2,5 @@ pub mod run;
 
 #[derive(Debug)]
 pub enum Action {
-    Run { pw_length: usize, num_pw: usize },
+    Run { pw_length: u8, num_pw: u8 },
 }
