+++ src/cli/dispatch/mod.rs
@@ -3,8 +3,8 @@ use anyhow::Result;
 
 pub fn handler(matches: &clap::ArgMatches) -> Result<Action> {
     Ok(Action::Run {
-        pw_length: matches.get_one::<usize>("pw_length").copied().unwrap_or(18),
-        num_pw: matches.get_one::<usize>("num_pw").copied().unwrap_or(1),
+        pw_length: matches.get_one::<u8>("pw_length").copied().unwrap_or(18),
+        num_pw: matches.get_one::<u8>("num_pw").copied().unwrap_or(1),
     })
 }
 
@@ -21,9 +21,9 @@ mod tests {
 
         let m = matches.unwrap();
 
-        assert_eq!(m.get_one::<usize>("pw_length").copied(), Some(18));
+        assert_eq!(m.get_one::<u8>("pw_length").copied(), Some(18));
 
-        assert_eq!(m.get_one::<usize>("num_pw").copied(), Some(1));
+        assert_eq!(m.get_one::<u8>("num_pw").copied(), Some(1));
 
         let action = handler(&m)?;
